Milestone4/requirements.txt
transformers
accelerate
torch
streamlit
pyngrok
python-dotenv
PyJWT
vaderSentiment
Pillow
pandas
matplotlib
requests
fastapi
uvicorn
nest_asyncio
bitsandbytes
sentencepiece
